v4v : buildable copy of version 9 for xen as submitted to xen-devel
Applies to Xen 4.3.4.

Almost identical to that submitted to upstream: does not address
the feedback received in response to that patch series.

As per original v9, viptables is renamed (some comments still need
to be updated) and not separated into its own patch.

xsm bits from OpenXT are isolated into a separate patch.

xc-xt-unmap-shared-info needed a tweak to apply afterwards.

Signed-off-by: Christopher Clark <christopher.w.clark@gmail.com>