OXT-1503: [xen] [argo] [xsm] transmit XSM sid in message header
Convey the peer Flask label of a communicating domain by introducing
the 32-bit Flask SID in the xen_argo_ring_message_header that is
delivered with each Argo message.

To retain the existing size of the message header while adding this
32-bit field the message_type field is resized from 32 to 8 bits and
a 16-bit padding field is removed. 8-bits of padding remain.

This is an ABI change and a corresponding update to the Linux Argo
device driver is required.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>