OXT-1503: Linux Argo support for xmit of XSM/Flask SID in msg header
Updates the Linux device driver to a revised Argo ABI where the
XSM/Flask SID of the sender domain is conveyed with each message.

The ring message header has been changed to accommodate the new data and
the message_type field is reduced in size to 8-bits. Stream and datagram
protocol indicator values are redefined to within the 8-bit range.

Works in conjunction with the corresponding patch for Xen.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>