distro conf : xenstored preference and multi-provider whitelist
Xenstored is now a virtual package and we select the C version which
is produced by the xen-tools recipe.

  PREFERRED_PROVIDER_virtual/xenstored = "xen-tools"
  PREFERRED_RPROVIDER_virtual/xenstored = "xen-tools-xenstored"

Both the xen-tools and xen-ocaml-libs recipes produce packages that
provide virtual/xenstored, which is a condition that triggers a build
error by default. The xenstored packages are implemented to be able to
coexist, so whitelisting virtual/xenstored as allowing multiple
providers is appropriate and removes the build error.

OXT-1694

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>